requirements.txt
Scrapy 
DateRangeParser 
python-dateutil 
bs4 
requests